+++ src/translations/it_IT.js
@@ -34,6 +34,7 @@ module.exports = {
   "hostedFieldsTokenizationCvvVerificationFailedError" : "Verifica della carta di credito non riuscita. Controlla i tuoi dati e riprova.",
   "paypalButtonMustBeUsed" : "Usa il pulsante PayPal per continuare con il pagamento.",
   "venmoButtonMustBeUsed": "Usa il pulsante Venmo per continuare il pagamento.",
+  "applePayButtonMustBeUsed": "Usa il pulsante Apple Pay per continuare con il pagamento.",
   "paypalAccountTokenizationFailedError" : "Si è verificato un errore durante l'aggiunta del conto PayPal. Riprova.",
   "paypalFlowFailedError" : "Si è verificato un errore durante la connessione a PayPal. Riprova.",
   "paypalTokenizationRequestActiveError" : "Autorizzazione di pagamento PayPal già in corso.",